feat: remove upstash (except ratelimit stuff), pnpm update --latest, fix plaiceholder
