+++ .env.example
@@ -5,8 +5,5 @@ REPLICATE_API_TOKEN=
 UPSTASH_REDIS_REST_URL=
 UPSTASH_REDIS_REST_TOKEN=
 
-# Follow the steps here to get your cloudflare worker URL (e.g. https://images.extrapolate.workers.dev)
-NEXT_PUBLIC_CLOUDFLARE_WORKER=
-
-# Generate a random token here: https://generate-secret.vercel.app/32
-CLOUDFLARE_WORKER_SECRET=
+# Cloudflare tunnel
+TUNNEL_URL=
