+++ package.json
@@ -8,51 +8,53 @@
     "format:write": "prettier --write \"**/*.{css,js,json,jsx,ts,tsx}\"",
     "format": "prettier \"**/*.{css,js,json,jsx,ts,tsx}\"",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "tunnel": "cloudflared tunnel --url http://localhost:3000"
   },
   "dependencies": {
-    "@radix-ui/react-popover": "^1.0.6",
+    "@radix-ui/react-popover": "^1.0.7",
     "@radix-ui/react-switch": "^1.0.3",
-    "@radix-ui/react-tooltip": "^1.0.6",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@supabase/ssr": "^0.3.0",
     "@supabase/supabase-js": "^2.43.1",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
-    "@upstash/ratelimit": "^0.4.4",
-    "@upstash/redis": "^1.22.0",
-    "@vercel/analytics": "^0.1.11",
-    "@vercel/og": "^0.0.26",
-    "classnames": "^2.3.2",
-    "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
-    "focus-trap-react": "^10.2.1",
-    "framer-motion": "^10.16.4",
-    "lucide-react": "0.105.0-alpha.4",
+    "@types/node": "20.12.10",
+    "@types/react": "18.3.1",
+    "@types/react-dom": "18.3.0",
+    "@upstash/ratelimit": "^1.1.3",
+    "@upstash/redis": "^1.30.1",
+    "@vercel/analytics": "^1.2.2",
+    "@vercel/og": "^0.6.2",
+    "classnames": "^2.5.1",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.3",
+    "focus-trap-react": "^10.2.3",
+    "framer-motion": "^11.1.9",
+    "lucide-react": "0.378.0",
     "ms": "^2.1.3",
-    "nanoid": "^4.0.2",
-    "next": "13.3.0",
-    "plaiceholder": "^2.5.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "nanoid": "^5.0.7",
+    "next": "14.2.3",
+    "plaiceholder": "^3.0.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-hot-toast": "^2.4.1",
-    "react-markdown": "^8.0.7",
-    "react-wrap-balancer": "^0.3.0",
-    "replicate": "^0.9.3",
-    "sharp": "^0.31.3",
-    "swr": "^2.2.2",
-    "typescript": "4.9.4",
-    "use-debounce": "^9.0.4"
+    "react-markdown": "^9.0.1",
+    "react-wrap-balancer": "^1.1.0",
+    "replicate": "^0.29.4",
+    "sharp": "^0.33.3",
+    "swr": "^2.2.5",
+    "typescript": "5.4.5",
+    "use-debounce": "^10.0.0"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.5.6",
+    "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/line-clamp": "^0.4.4",
-    "@tailwindcss/typography": "^0.5.10",
-    "@types/ms": "^0.7.31",
-    "autoprefixer": "^10.4.16",
-    "concurrently": "^7.6.0",
-    "postcss": "^8.4.30",
-    "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.2.8",
-    "tailwindcss": "^3.3.3"
+    "@tailwindcss/typography": "^0.5.13",
+    "@types/ms": "^0.7.34",
+    "autoprefixer": "^10.4.19",
+    "concurrently": "^8.2.2",
+    "postcss": "^8.4.38",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.14",
+    "tailwindcss": "^3.4.3"
   }
 }
