+++ pages/api/images/[id]/index.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
-import { getData } from "@/lib/upstash";
+import { supabase } from "@/lib/supabase";
+import { DataProps } from "@/lib/types";
 
 export const config = {
   runtime: "edge",
@@ -7,6 +8,12 @@ export const config = {
 
 export default async function handler(req: NextRequest) {
   const id = req.nextUrl.pathname.split("/")[3];
-  const data = await getData(id);
+  // TODO: error handling
+  const { data, error } = await supabase
+    .from("data")
+    .select("*")
+    .eq("id", id)
+    .returns<DataProps[]>()
+    .single();
   return NextResponse.json(data);
 }
